refactor (create CopyWithOneLink-component) and  clean up code
