+++ components/SnippetDetails.js
@@ -2,8 +2,7 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import styled from "styled-components";
 import useSWR from "swr";
-import { useState } from "react";
-import { Icon } from "@iconify/react";
+import CopyWithOneClick from "./CopyWithOneClick";
 
 function SnippetDetails({ onDelete }) {
   const router = useRouter();
@@ -11,15 +10,6 @@ function SnippetDetails({ onDelete }) {
   const { id } = router.query;
 
   const { data, isLoading, error } = useSWR(`/api/snippets/${id}`);
-
-  const [isCopied, setIsCopied] = useState(false);
-  async function handleClick() {
-    setIsCopied(true);
-    setTimeout(() => {
-      setIsCopied(false);
-    }, 3000);
-    await navigator.clipboard.writeText(data.code);
-  }
 
   if (error) {
     return <div>Failed to load Details View 🥺</div>;
@@ -36,15 +26,8 @@ function SnippetDetails({ onDelete }) {
       <Heading>Code</Heading>
       <CodeContainer>
         {code}
-        <StyledButton type="button" onClick={handleClick}>
-          <Icon
-            icon={isCopied === true ? "mingcute:check-fill" : "fa-regular:copy"}
-            aria-label={isCopied === true ? "code copied" : "copy code"}
-          />
-          {isCopied === true ? "code copied" : "copy code"}
-        </StyledButton>
+        <StyledCopyWithOneClick hasIconText codeData={code} />
       </CodeContainer>
-
       <Heading>Description</Heading>
       <p>{description}</p>
       <Heading>Further Resources</Heading>
@@ -87,15 +70,10 @@ const Heading = styled.h3`
   color: var(--primary-color);
 `;
 
-const StyledButton = styled.button`
+const StyledCopyWithOneClick = styled(CopyWithOneClick)`
   position: absolute;
-  background: transparent;
-  border: none;
   right: 0.25rem;
   top: 0.25rem;
-  display: flex;
-  gap: 0.3rem;
-  align-items: center;
 `;
 
 const CodeContainer = styled.div`
